+++ CHANGELOG.rst
@@ -1,4 +1,7 @@
-`Unreleased <https://github.com/pace-neutrons/Euphonic/compare/v0.3.1...HEAD>`_
+`Unreleased <https://github.com/pace-neutrons/Euphonic/compare/v0.3.2...HEAD>`_
+----------
+
+`v0.3.2 <https://github.com/pace-neutrons/Euphonic/compare/v0.3.1...v0.3.2>`_
 ----------
 
 - New Features:
@@ -20,6 +23,7 @@
       data as 1D spectrum objects. (It is anticipated that this module
       will grow to include schemes beyond this average over a single
       sphere.)
+  - Added ``Crystal.to_spglib_cell`` convenience function
 
 - Changes:
 
@@ -30,6 +34,9 @@
     computer. (e.g. euphonic-dos)
   - From an interactive shell with tab-completion, one can find all
     the euphonic tools by typing "euphonic-<TAB>".
+  - Changed arguments for ``util.get_qpoint_labels(Crystal, qpts)``
+    to ``util.get_qpoint_labels(qpts, cell=None)`` where
+    ``cell = Crystal.to_spglib_cell()``
 
 - Bug fixes: